move `unescape` and `dedent` into `util` module
